requirements.txt
attrs==20.3.0
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.6
cssselect==1.1.0
elasticsearch==7.12.0
elasticsearch-dsl==7.3.0
greenlet==1.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
parsel==1.6.0
Pillow==8.2.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
pyTelegramBotAPI==3.7.6
python-dateutil==2.8.1
queuelib==1.5.0
requests==2.25.1
Scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
SQLAlchemy==1.4.4
Twisted==21.2.0
twisted-iocpsupport==1.0.1
urllib3==1.26.4
w3lib==1.22.0
zope.interface==5.3.0
